- Refator main js - Move some codes to util js
